+++ CHANGELOG.md
@@ -31,6 +31,7 @@
   - Deprecated `receiverhelper.WithLogs` in favour of `component.WithLogsReceiver`
   - Deprecated `receiverhelper.NewFactory` in favour of `component.NewReceiverFactory`
 - Remove support for legacy otlp/http port. (#4916)
+- Remove deprecated funcs in pdata (#4809)
 
 ### 💡 Enhancements 💡